+++ CHANGELOG.md
@@ -6,11 +6,14 @@ Changelog
 
 ### Features
 
+- Start and stop Mantis Client from Daedalus main process ([PR 568](https://github.com/input-output-hk/daedalus/pull/568))
+
 ### Fixes
 
 ### Chores
 
 - Update DLL package and cleanup translation files ([PR 566](https://github.com/input-output-hk/daedalus/pull/566))
+- Replace electron-json-storage with electron-storage ([PR 579](https://github.com/input-output-hk/daedalus/pull/579))
 
 ## 0.8.3
 =======